site_agent_tg/requirements-bot.txt
aiogram==3.4
aiohttp
python-dotenv